requirements.txt
pandas>=1.5.0
scikit-learn>=1.0.0
streamlit>=1.28.0
requests>=2.28.0
numpy>=1.21.0
python-dotenv>=0.21.0
